+++ tmp_export/events_post.py
@@ -1,0 +1,4 @@
+import math
+
+def events_post(t, y_vec, params, evt_log_scratch, runtime_ws):
+    return (-1, 0)